user_following permissions checked

+++ user_profile/views/user_following.py
@@ -29,10 +29,17 @@ class UserFollowingViewSet(viewsets.GenericViewSet):
 
     def create(self, request, following_user__username):
         username = following_user__username
-        resource_user = self.queryset.get(user__username=username)
-        # self.check_object_permissions(request, resource_user)
+
         if "username" not in request.data:
             raise exceptions.ValidationError({"username": "Required field"})
+        try:
+            resource_user = self.queryset.get(user__username=username)
+            self.check_object_permissions(request, resource_user)
+        except ObjectDoesNotExist:
+            raise exceptions.ValidationError(
+                {"username": "The username provided did not match any know users."}
+            )
+
         username_to_follow = request.data["username"]
         try:
             user_to_follow = self.queryset.get(user__username=username_to_follow)
@@ -41,7 +48,9 @@ class UserFollowingViewSet(viewsets.GenericViewSet):
             )
         except ObjectDoesNotExist:
             raise exceptions.ValidationError(
-                {"username": "The username provided did not match any know users."}
+                {
+                    "username": "The username provided for the account to follow did not match any know users."
+                }
             )
 
         except IntegrityError:
@@ -60,6 +69,7 @@ class UserFollowingViewSet(viewsets.GenericViewSet):
         username = following_user__username
         try:
             resource_user = self.queryset.get(user__username=username)
+            self.check_object_permissions(request, resource_user)
             user_to_unfollow = self.queryset.get(user__username=unfollow_username)
 
             resource_user.following.get(user_profile=user_to_unfollow).delete()
